+++ tests/src/containers/freelist_tests.c
@@ -48,7 +48,7 @@ u8 freelist_should_allocate_one_and_free_one() {
     freelist_create(total_size, &memory_requirement, block, &list);
 
     // allocate some space
-    u32 offset = INVALID_ID;  // start with invalid id, which is a good default since it should never happen
+    u64 offset = INVALID_ID;  // start with invalid id, which is a good default since it should never happen
     b8 result = freelist_allocate_block(&list, 64, &offset);
     // verify that result is true, offset should be set to 0
     expect_to_be_true(result);
@@ -88,21 +88,21 @@ u8 freelist_should_allocate_one_and_free_multi() {
     freelist_create(total_size, &memory_requirement, block, &list);
 
     // allocate some space
-    u32 offset = INVALID_ID;  // start with an invalid id, which is a good default since it should never happen
+    u64 offset = INVALID_ID;  // start with an invalid id, which is a good default since it should never happen
     b8 result = freelist_allocate_block(&list, 64, &offset);
     // verify taht the result is true, offset should be set to 0
     expect_to_be_true(result);
     expect_should_be(0, offset);
 
     // allocate some more space
-    u32 offset2 = INVALID_ID;  // start with an invalid id, which is a good default since it should never happen
+    u64 offset2 = INVALID_ID;  // start with an invalid id, which is a good default since it should never happen
     result = freelist_allocate_block(&list, 64, &offset2);
     // verify that result is true, offset should be set to the size of the previous allocation
     expect_to_be_true(result);
     expect_should_be(64, offset2);
 
     // allocate some more space
-    u32 offset3 = INVALID_ID;  // start with an invalid id, which is a good default since it should never happen
+    u64 offset3 = INVALID_ID;  // start with an invalid id, which is a good default since it should never happen
     result = freelist_allocate_block(&list, 64, &offset3);
     // verify that result is true, offset should be set to the size of the previous allocation
     expect_to_be_true(result);
@@ -122,7 +122,7 @@ u8 freelist_should_allocate_one_and_free_multi() {
     expect_should_be(total_size - 128, free_space);
 
     // allocate some more space - this should fill the middle block back in
-    u32 offset4 = INVALID_ID;  // start with an invalid id, which is a good default since it should never happen
+    u64 offset4 = INVALID_ID;  // start with an invalid id, which is a good default since it should never happen
     result = freelist_allocate_block(&list, 64, &offset4);
     // verify that result is true, offset should be set to the size of the previous allocation
     expect_to_be_true(result);
@@ -171,21 +171,21 @@ u8 freelist_should_allocate_one_and_free_multi_varying_sizes() {
     freelist_create(total_size, &memory_requirement, block, &list);
 
     // allocate some space
-    u32 offset = INVALID_ID;  // start with an invalid id, which is a good default since it should never happen
+    u64 offset = INVALID_ID;  // start with an invalid id, which is a good default since it should never happen
     b8 result = freelist_allocate_block(&list, 64, &offset);
     // verify taht the result is true, offset should be set to 0
     expect_to_be_true(result);
     expect_should_be(0, offset);
 
     // allocate some more space
-    u32 offset2 = INVALID_ID;  // start with an invalid id, which is a good default since it should never happen
+    u64 offset2 = INVALID_ID;  // start with an invalid id, which is a good default since it should never happen
     result = freelist_allocate_block(&list, 32, &offset2);
     // verify that result is true, offset should be set to the size of the previous allocation
     expect_to_be_true(result);
     expect_should_be(64, offset2);
 
     // allocate some more space
-    u32 offset3 = INVALID_ID;  // start with an invalid id, which is a good default since it should never happen
+    u64 offset3 = INVALID_ID;  // start with an invalid id, which is a good default since it should never happen
     result = freelist_allocate_block(&list, 64, &offset3);
     // verify that result is true, offset should be set to the size of the previous allocation
     expect_to_be_true(result);
@@ -205,7 +205,7 @@ u8 freelist_should_allocate_one_and_free_multi_varying_sizes() {
     expect_should_be(total_size - 128, free_space);
 
     // allocate some more space, this time larger than the old middle block. this should have a new offset at the end of the list
-    u32 offset4 = INVALID_ID;  // start with an invalid id, which is a good default since it should never happen
+    u64 offset4 = INVALID_ID;  // start with an invalid id, which is a good default since it should never happen
     result = freelist_allocate_block(&list, 64, &offset4);
     // verify that result is true, offset should be set to the size of the previous allocation
     expect_to_be_true(result);
@@ -254,7 +254,7 @@ u8 freelist_should_allocate_to_full_and_fail_to_allocate_more() {
     freelist_create(total_size, &memory_requirement, block, &list);
 
     // allocate all the space
-    u32 offset = INVALID_ID;  // start with invalid id, which is a good default since it should never happen
+    u64 offset = INVALID_ID;  // start with invalid id, which is a good default since it should never happen
     b8 result = freelist_allocate_block(&list, 512, &offset);
     // verify that result is true, offset should be set to 0
     expect_to_be_true(result);
@@ -265,7 +265,7 @@ u8 freelist_should_allocate_to_full_and_fail_to_allocate_more() {
     expect_should_be(0, free_space);
 
     // now try allocating some more
-    u32 offset2 = INVALID_ID;
+    u64 offset2 = INVALID_ID;
     KDEBUG("The following warning message is intentional.");
     result = freelist_allocate_block(&list, 64, &offset2);
     // verify that result is false
